feat(site): merge Limits + Usage into unified Spend page
