+++ site/src/pages/AgentsPage/components/LimitsTab/UserOverridesSection.tsx
@@ -23,6 +23,7 @@ import { ConfirmDeleteDialog } from "../ConfirmDeleteDialog";
 import { SectionHeader } from "../SectionHeader";
 
 interface UserOverridesSectionProps {
+	hideHeader?: boolean;
 	overrides: ReadonlyArray<{
 		user_id: string;
 		name: string;
@@ -55,6 +56,7 @@ interface UserOverridesSectionProps {
 }
 
 export const UserOverridesSection: FC<UserOverridesSectionProps> = ({
+	hideHeader,
 	overrides,
 	showUserForm,
 	onShowUserFormChange,
@@ -80,10 +82,12 @@ export const UserOverridesSection: FC<UserOverridesSectionProps> = ({
 
 	return (
 		<section className="space-y-4">
-			<SectionHeader
-				label="Per-User Overrides"
-				description="Override the deployment default spend limit for specific users. User overrides take highest priority, followed by group limits, then the deployment default."
-			/>
+			{!hideHeader && (
+				<SectionHeader
+					label="Per-User Overrides"
+					description="Override the deployment default spend limit for specific users. User overrides take highest priority, followed by group limits, then the deployment default."
+				/>
+			)}
 			<div className="space-y-4">
 				{overrides.length > 0 ? (
 					<Table>
